refactor: move filter_state_according_to_player function to utils and update imports
