+++ langchain_werewolf/game_players/__init__.py
@@ -7,6 +7,7 @@ from .const import (
 )
 from .registry import PlayerSideRegistry, PlayerRoleRegistry
 from .utils import (
+    filter_state_according_to_player,
     find_player_by_name,
     find_players_by_role,
     find_players_by_side,
@@ -23,6 +24,7 @@ __all__ = [
     BaseGamePlayerRole.__name__,
     PlayerSideRegistry.__name__,
     PlayerRoleRegistry.__name__,
+    filter_state_according_to_player.__name__,
     find_player_by_name.__name__,
     find_players_by_role.__name__,
     find_players_by_side.__name__,
